Update page_size and cpu_core to be int32 Closes #10 (#654)

+++ src/TesApi.Web/Controllers/TaskServiceApi.cs
@@ -289,7 +289,7 @@ namespace TesApi.Controllers
         [ValidateModelState]
         [SwaggerOperation("ListTasks")]
         [SwaggerResponse(statusCode: 200, type: typeof(TesListTasksResponse), description: "")]
-        public virtual async Task<IActionResult> ListTasks([FromQuery] string namePrefix, [FromQuery] long? pageSize, [FromQuery] string pageToken, [FromQuery] string view, CancellationToken cancellationToken)
+        public virtual async Task<IActionResult> ListTasks([FromQuery] string namePrefix, [FromQuery] int? pageSize, [FromQuery] string pageToken, [FromQuery] string view, CancellationToken cancellationToken)
         {
             var decodedPageToken =
                 pageToken is not null ? Encoding.UTF8.GetString(Base64UrlTextEncoder.Decode(pageToken)) : null;
